More accuracy ScenarioNavigatorItemDelegate moving
